Code refactor + add recipe reader

+++ app/src/main/java/com/example/bookofrecipes/dataBase/AppDatabaseEntity.kt
@@ -13,7 +13,7 @@ data class Ingredient(
 data class Recipe(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String,
-    val type: Int,
+    val type: Long,
     val recipeText: String,
     val isFavorite: Boolean
 )
